for issue 94 - TF loads when loading TorchClient (after comments)
