+++ grid/clients/base.py
@@ -194,62 +194,4 @@ class BaseClient(base_worker.GridWorker):
         data = json.dumps([{'name': name, 'address': addr}])
         self.publish('openmined:add_task', data)
 
-    def best_models(self, task):
-        self.show_models = widgets.VBox([widgets.HBox([widgets.Label('Model Address')])])
-        self.listen_to_channel(channels.add_model(task), self.__added_model)
-        self.publish(channels.list_models, task)
-
-        return self.show_models
-
-    def __added_model(self, message):
-        info = self.api.get_json(message['data'])
-        model_addr = info['model']
-
-        hbox = widgets.HBox([widgets.Label(model_addr)])
-        self.show_models.children += (hbox,)
-
-    def load_model(self, addr):
-        return utils.ipfs2keras(addr)
-
-    def send_model(self, name, model_addr):
-        task = utils.load_task(name)
-
-        update = {
-            'name': name,
-            'model': model_addr,
-            'task': task['address'],
-            'creator': self.id,
-            'parent': task['address']
-        }
-
-        update_addr = self.api.add_json(update)
-        self.publish(channels.add_model(name), update_addr)
-
-        print("SENDING MODEL!!!!")
-
-    def add_model(self, name, model, parent=None):
-        """
-        Propose a model as a solution to a task.
-        parent  - The name of the task.  e.g. MNIST
-        model - A keras model. Down the road we should support more frameworks.
-        """
-        task = utils.load_task(name)
-        p = None
-        if parent is None:
-            p = task['address']
-        else:
-            p = parent
-
-        model_addr = utils.keras2ipfs(model)
-
-        update = {
-            'name': name,
-            'model': model_addr,
-            'task': task['address'],
-            'creator': self.id,
-            'parent': p
-        }
-
-        update_addr = self.api.add_json(update)
-        self.publish(channels.add_model(name), update_addr)
-        print(f"ADDED NEW MODELS WEIGHT TO {update_addr}")
+    """ the following moved to clients/keras.py """
